# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyBank API",
    "description": "API per progetto universitario bancario MyBank sviluppato con Next.js",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/"
    }
  ],
...
  "paths": {
...
    "/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Ottieni lista conti utente autenticato",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista conti utente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "account_number": {
                            "type": "string"
                          },
                          "account_type": {
                            "type": "string",
                            "enum": ["savings", "checking"]
                          },
                          "balance": {
                            "type": "number",
                            "format": "float"
                          },
                          "status": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID account non valido"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore recupero conti"
          }
        }
      }
    },
    "/accounts/new": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Crea nuovo conto bancario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account_type"],
                "properties": {
                  "account_type": {
                    "type": "string",
                    "enum": ["savings", "checking"],
                    "example": "savings"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conto creato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "account_number": {
                          "type": "string"
                        },
                        "account_type": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "number",
                          "format": "float"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tipo conto non valido o conto duplicato"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore nella creazione conto"
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Dettagli di un account specifico con statistiche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID univoco dell'account"
          }
        ],
        "responses": {
          "200": {
            "description": "Dettagli dell'account con statistiche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "account_number": {
                          "type": "string"
                        },
                        "account_type": {
                          "type": "string",
                          "enum": ["savings", "checking"]
                        },
                        "balance": {
                          "type": "number",
                          "format": "float"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "statistics": {
                          "type": "object",
                          "properties": {
                            "transaction_count": {
                              "type": "integer"
                            },
                            "last_transaction_date": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true
                            },
                            "monthly_income": {
                              "type": "number",
                              "format": "float"
                            },
                            "monthly_expenses": {
                              "type": "number",
                              "format": "float"
                            },
                            "monthly_transactions": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID account non valido"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "404": {
            "description": "Conto non trovato"
          },
          "500": {
            "description": "Errore nel recupero conto"
          }
        }
      }
    },
...
    "/transactions/deposit": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Esegue un deposito su un conto bancario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account_id", "amount"],
                "properties": {
                  "account_id": {
                    "type": "integer",
                    "example": 123,
                    "description": "Id del conto su cui effettuare il deposito"
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "example": 100.5,
                    "description": "Importo del deposito"
                  },
                  "description": {
                    "type": "string",
                    "example": "Deposito",
                    "description": "Descrizione opzionale del deposito"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposito completato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "account_id": {
                          "type": "integer"
                        },
                        "amount": {
                          "type": "number",
                          "format": "float"
                        },
                        "transaction_type": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione input"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore nel processamento del deposito"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}